creation+insertion done

+++ app/src/main/java/com/example/rishabhtoys/DetailEntityViewModel.kt
@@ -9,7 +9,7 @@ class DetailEntityViewModel(application: Application) : AndroidViewModel(applica
     var repository: Repository = Repository(application)
 
 
-    fun getDetailEntityInfo(entityId: Int?) {
+    fun getDetailEntityInfo(entityId: Long?) {
         repository.getDetailInfoForEntity(entityId!!)
     }
